pyproject.toml
[tool.poetry]
name = "kinconnect-api"
version = "0.1.0"
description = ""
authors = ["Nehil Jain"]
readme = "README.md"

packages = [
    { include = "kinconnect_api" }
]

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.2.1"
ipykernel = "^6.29.4"
ipdb = "^0.13.13"
pydantic = "^2.7.3"
langsmith = "^0.1.67"
thefuzz = "^0.22.1"
pandas = "^2.2.2"
groq = "^0.8.0"
requests = "^2.32.3"
langchain = {extras = ["all"], version = "^0.2.2"}
langchain-core = "^0.2.4"
langchain-community = "^0.2.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
langchain-mongodb = "^0.1.6"
pymongo = "^4.7.3"
langchain-fireworks = "^0.1.3"
modal = "^0.62.223"
langchain-experimental = "^0.0.61"
langchain-text-splitters = "^0.2.2"
motor = "^3.5.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff.lint.isort]
known_first_party = ["kinconnect_api"]
force_sort_within_sections = true
